+++ backend/interfaces/api/server/handler/index.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 
 	"github.com/gorilla/mux"
-	"github.com/yoshihiro-shu/draft-backend/request"
+	"github.com/yoshihiro-shu/draft-backend/interfaces/api/server/request"
 )
 
 type Handler struct {
